simplify runner
